Add metadata to Scramble page
